+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Implement FTS table for `PatientAddress`
 - Define overdue appointment query as a `const` in `OverdueAppointment`
 - Bump Annotation library to v1.4.0
+- Provide overdue search history string set from preferences
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-06-13-8291